Cast totals to int so that the field comes back as a number, not a bigint

+++ server/lib/models/goal/get-target-totals.sql
@@ -33,7 +33,7 @@ tallies AS (
 )
 SELECT
 	targets."goalId",
-	sum(tallies.count) AS total
+	sum(tallies.count)::int AS total
 FROM
 	targets
 	INNER JOIN tallies ON (
